pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pass_culture"
version = "0.1.0"
description = "An async API Wrapper for the Pass Culture"
authors = [{ name = "Esteban" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.8"
dependencies = ["httpx>=0.23.0", "pydantic>=1.9.0", "python-dotenv>=0.20.0"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.18.0",
    "black>=22.3.0",
    "isort>=5.10.0",
]
